Put Kafka Harvester and HWC Streaming Harvester together
