--- spark-atlas-connector/src/main/scala/org/apache/spark/sql/kafka010/atlas/KafkaTopicInformation.scala
@@ -1,27 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.kafka010.atlas
-
-case class KafkaTopicInformation(topicName: String, clusterName: Option[String] = None)
-
-object KafkaTopicInformation {
-  def getQualifiedName(ti: KafkaTopicInformation, defaultClusterName: String): String = {
-    val cName = ti.clusterName.getOrElse(defaultClusterName)
-    s"${ti.topicName}@$cName"
-  }
-}
